requirements.txt
absl-py     ==              2.2.2
altair        ==            5.5.0
asttokens       ==          3.0.0
astunparse        ==        1.6.3
attrs               ==      25.3.0
blinker               ==    1.9.0
cachetools              ==  5.5.2
certifi                   ==2025.4.26
charset-normalizer  ==      3.4.2
click                 ==    8.2.1
colorama                ==  0.4.6
comm                      ==0.2.2
contourpy==                 1.3.2
cycler     ==               0.12.1
debugpy      ==             1.8.14
decorator      ==           5.2.1
easyocr          ==         1.7.1
executing          ==       2.2.0
filelock             ==     3.18.0
flatbuffers            ==   25.2.10
fonttools                == 4.58.0
fsspec         ==           2025.5.0
gast             ==         0.6.0
gitdb              ==       4.0.12
GitPython            ==     3.1.44
google-pasta           ==   0.2.0
grpcio                   == 1.71.0
h5py==                      3.13.0
idna  ==                    3.10
imageio ==                  2.37.0
ipykernel ==                6.29.5

ipython_pygments_lexers==   1.1.1
jedi   ==                   0.19.2
Jinja2   ==                 3.1.6
jsonschema ==               4.23.0
jsonschema-specifications== 2025.4.1
jupyter_client==            8.6.3
jupyter_core    ==          5.7.2
keras             ==        3.10.0
kiwisolver          ==      1.4.8
lazy_loader           ==    0.4
libclang                ==  18.1.1
Markdown                  ==3.8
markdown-it-py ==           3.0.0
MarkupSafe       ==         3.0.2
matplotlib         ==       3.10.3
matplotlib-inline    ==     0.1.7
mdurl                  ==   0.1.2
ml-dtypes                == 0.4.1
mpmath   ==                 1.3.0
namex      ==               0.0.9
narwhals     ==             1.40.0
nest-asyncio   ==           1.6.0
networkx         ==         3.4.2
ninja              ==       1.11.1.4
numpy                ==     1.26.4
opencv-python-headless ==   4.9.0.80
opt_einsum  ==              3.4.0
optree        ==            0.15.0
packaging       ==          24.2
pandas            ==        2.2.3
parso               ==      0.8.4
Pillow                ==    9.5.0
pip                     ==  24.3.1
platformdirs  ==            4.3.8
prompt_toolkit  ==          3.0.51
protobuf          ==        4.25.7
psutil              ==      7.0.0
pure_eval             ==    0.2.3
pyarrow                 ==  20.0.0
pyclipper                 ==1.3.0.post6
pydeck     ==               0.9.1
Pygments     ==             2.19.1
pyparsing      ==           3.2.3
python-bidi      ==         0.6.6
python-dateutil    ==       2.9.0.post0
pytz                 ==     2025.2
pywin32                ==   310
PyYAML                   == 6.0.2
pyzmq   ==                  26.4.0
referencing==               0.36.2
requests     ==             2.32.3
rich           ==           13.9.4
rpds-py          ==         0.25.1
scikit-image       ==       0.24.0
scipy                ==     1.15.3
setuptools             ==   80.8.0
shapely                  == 2.1.1
six  ==                     1.17.0
smmap  ==                   5.0.2
stack-data==                0.6.3
streamlit~=1.45.1
sympy         ==            1.14.0
tenacity        ==          8.5.0
tensorboard       ==        2.17.1
tensorboard-data-server==   0.7.2
tensorflow               == 2.17.0
tensorflow-intel==          2.17.0
termcolor         ==        3.1.0
tifffile            ==      2025.5.10
toml                  ==    0.10.2
torch                   ==  2.7.0
torchvision               ==0.22.0
tornado    ==               6.5
traitlets    ==             5.14.3
typing_extensions==         4.13.2
tzdata             ==       2025.2
urllib3              ==     2.4.0
watchdog               ==   6.0.0
wcwidth                  == 0.2.13
Werkzeug ==                 3.1.3
wheel   ==                  0.45.1
wrapt     ==                1.17.2